Move some methods into traits
